+++ lib/objects/voucher_children/limited_percentage_voucher_without_end_time.dart
@@ -55,7 +55,7 @@ class LimitedPercentageVoucherWithoutEndTime extends Voucher
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    int? discountValue,
+    double? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
