+++ lib/objects/voucher_children/unlimited_amount_voucher_without_end_time.dart
@@ -28,7 +28,7 @@ class UnlimitedAmountVoucherWithoutEndTime extends Voucher {
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    int? discountValue,
+    double? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
